Provide correct search.

+++ QBuilder/QBuilder/src/createsecurityfilterform.cpp
@@ -3,12 +3,15 @@
 
 #include <QSqlRecord>
 #include <QSqlQuery>
+#include <QMessageBox>
 
 CreateSecurityFilterForm::CreateSecurityFilterForm(QWidget *parent) :
     BaseForm(parent),
     ui(new Ui::CreateSecurityFilterForm)
 {
     ui->setupUi(this);
+
+    ui->submitButton->setEnabled(false); // All columns
 
     initData(Table);
 
@@ -27,7 +30,7 @@ CreateSecurityFilterForm::CreateSecurityFilterForm(QWidget *parent) :
     model->insertRow(model->rowCount(QModelIndex()));
     mapper->toLast();
 
-    questionModel = new BaseComboModel("type", "question", this, Table, "");
+    questionModel = new BaseComboModel("entry_name || ': ' || question", "security_question", this, Table, "");
     ui->questionListView->setModel(questionModel);
     ui->questionListView->setRowHidden(0, true);
     ui->questionListView->setSelectionMode(QAbstractItemView::MultiSelection);
@@ -51,6 +54,19 @@ CreateSecurityFilterForm::CreateSecurityFilterForm(QWidget *parent) :
     });
 
     connect(ui->submitButton, &QToolButton::clicked, this, &CreateSecurityFilterForm::submitChanges);
+
+    // All columns
+    connect(ui->userTypeLine, &QLineEdit::textChanged, this, [this] {
+        ui->submitButton->setEnabled(!ui->userTypeLine->text().isEmpty() && ui->accountTypeBox->currentIndex() > 0 &&
+        ui->tenantBox->currentIndex() > 0 && !ui->filter_1_Line->text().isEmpty() && !ui->filter_2_Line->text().isEmpty() &&
+        ui->regScreenBox->currentIndex() > 0);
+    });
+    connect(ui->accountTypeBox, &QComboBox::currentTextChanged, ui->userTypeLine, &QLineEdit::textChanged);
+    connect(ui->tenantBox, &QComboBox::currentTextChanged, ui->userTypeLine, &QLineEdit::textChanged);
+
+    connect(ui->filter_1_Line, &QLineEdit::textChanged, ui->userTypeLine, &QLineEdit::textChanged);
+    connect(ui->filter_2_Line, &QLineEdit::textChanged, ui->userTypeLine, &QLineEdit::textChanged);
+    connect(ui->regScreenBox, &QComboBox::currentTextChanged, ui->userTypeLine, &QLineEdit::textChanged);
 }
 
 CreateSecurityFilterForm::~CreateSecurityFilterForm()
@@ -62,37 +78,52 @@ void CreateSecurityFilterForm::submitChanges()
 {
     // Save changes to database
 
-    mapper->submit();
-    model->submitAll();
-
-    int id = -1;
-    if (isEdit) {
-        id = model->record(mapper->currentIndex()).value("id").toInt();
-    }
-
-    // Save data from comboBoxes to database
-    for (int i = 0; i < cbModel.count(); i++) {
-        cbModel.at(i)->saveToDB(combo.at(i)->itemData(combo.at(i)->currentIndex(), Qt::UserRole).toInt(), id);
-    }
-
-    // Save data from listViews
-    auto saveListData = [&id](QModelIndexList list, QString table, QString column)
-    {
-        for (QModelIndex index : list) {
-            int itemId = index.data(Qt::UserRole).toInt();
-            QString str = QString("INSERT INTO %1 (filter_id, %2) VALUES (%3, %4)").arg(table).arg(column).arg(id).arg(itemId);
-            QSqlQuery query;
-            query.exec(str);
-        }
-    };
-    QString str = QString("DELETE FROM filters_in_authorization WHERE filter_id = %1").arg(id);
     QSqlQuery query;
-    query.exec(str);
-    saveListData(ui->authTypeListView->selectionModel()->selectedIndexes(), "filters_in_authorization", "auth_id");
-    str = QString("DELETE FROM filters_in_question WHERE filter_id = %1").arg(id);
-    query.exec(str);
-    saveListData(ui->questionListView->selectionModel()->selectedIndexes(), "filters_in_question", "quest_id");
+    QString str = QString("SELECT EXISTS (SELECT " + Record + " FROM" + Table +
+                          " WHERE '" + Record + "' = '%1' AND id != %2 )").arg(ui->userTypeLine->text()).
+            arg(model->data(model->index(mapper->currentIndex(), 0), Qt::DisplayRole).toInt());
 
+    query.exec(str);
+    query.next();
+
+    // If exists
+    if (query.value(0) != 0 && !isEdit) {
+        QMessageBox::information(this, trUtf8("Error"), trUtf8("User type is already exists"));
+        return;
+    }
+    else {
+        // Insert new data
+        mapper->submit();
+        model->submitAll();
+
+        int id = -1;
+        if (isEdit) {
+            id = model->record(mapper->currentIndex()).value("id").toInt();
+        }
+
+        // Save data from comboBoxes to database
+        for (int i = 0; i < cbModel.count(); i++) {
+            cbModel.at(i)->saveToDB(combo.at(i)->itemData(combo.at(i)->currentIndex(), Qt::UserRole).toInt(), id);
+        }
+
+        // Save data from listViews
+        auto saveListData = [&id](QModelIndexList list, QString table, QString column) {
+            for (QModelIndex index : list) {
+                int itemId = index.data(Qt::UserRole).toInt();
+                QString str = QString("INSERT INTO %1 (filter_id, %2) VALUES (%3, %4)").arg(table).arg(column).arg(
+                        id).arg(itemId);
+                QSqlQuery query;
+                query.exec(str);
+            }
+        };
+        QString str = QString("DELETE FROM filters_in_authorization WHERE filter_id = %1").arg(id);
+        QSqlQuery query;
+        query.exec(str);
+        saveListData(ui->authTypeListView->selectionModel()->selectedIndexes(), "filters_in_authorization", "auth_id");
+        str = QString("DELETE FROM filters_in_question WHERE filter_id = %1").arg(id);
+        query.exec(str);
+        saveListData(ui->questionListView->selectionModel()->selectedIndexes(), "filters_in_question", "quest_id");
+    }
     model->select();
     mapper->toLast();
